+++ config/wdio/services.default-config.js
@@ -18,7 +18,7 @@ const serveStatic = {
 
 /* Selenium Docker Service Defaults */
 const seleniumDocker = {
-  /* Weather or not the service should be ran. */
+  /* Whether or not the service should be run. */
   enabled: true,
   /* Retry count to test for selenium being up. */
   retries: 4000,
@@ -26,7 +26,7 @@ const seleniumDocker = {
    */
   retryInterval: 10,
   /* Docker compose file reference when deploying the docker selenium hub stack. */
-  composeFile: path.join(__dirname, '..', '..', 'lib', 'wdio', 'docker-compose.yml'),
+  composeFile: path.join(__dirname, '..', 'docker', 'local-docker-compose.yml'),
 };
 
 module.exports = {